services/web/requirements.txt
Flask==1.1.2
Flask-AutoIndex==0.6.6
Flask-Silk==0.2
Pillow==7.1.2
gunicorn==20.0.4
